requirements.txt
django==3.0.7
djangorestframework==3.11.1
coreapi
